feat(sf-342a): relocate frontier contract to server-rust + finalize trait surface
R1 confirm/relocate decision: RELOCATE tombstone_frontier.rs core-rust -> server-rust.
Dependency-direction check (verified live): nothing in core-rust consumes the
contract; the sole implementer is server-side (epoch counter, per-client frontier,
reconnect gate). server-rust is the correct home; the wire tombstone shape stays
Vec<String> in core-rust.

Finalized trait/type surface (types/traits ONLY, no bodies):
- Add pub(crate) sub-trait CausalFrontierRanking: CausalFrontier carrying
  tracked_cursors() -> Vec<(ClientId, Epoch)>, kept OFF the general CausalFrontier
  so a bare &CausalFrontier holder cannot compute a subset-MIN prune path; the
  point-lookup client_cursor() is not on the surface. pub(crate) encodes
  'only the RAM-pressure forget-driver ranks' in the visibility system.
  Return shape is Vec (object-safe) so the downstream forget-driver may use
  &dyn dispatch.
- Declare PruneSafety: CausalFrontier (supertrait EXPOSES low_water_mark;
  fleet-wide-MIN stays doc-enforced/reviewer-checkable, not type-enforced).
- gate_decision_holds_at_commit consumes token by value; GateToken drops Clone
  (not Copy) for a genuine single-use/linear token. Re-admission safety rests on
  the commit-time re-check, not token unforgeability (impl-note for the gate child).
- GateToken fields kept pub with an accepted-forgeability doc note (forgeability
  is not load-bearing); avoids a constructor body that would violate types-only.
- confirm_apply/is_tracked doc-contracts state tracking is implicit via first
  confirm_apply, 'unknown == forgotten' collapses first-connect and
  forgotten-reconnect into the same gated path, and no track_client/register
  method exists or is needed.
- MaxRetention/GateToken remain serde-free (server-internal metadata).

cargo build/fmt/clippy (-D warnings) green for both crates.
